requirements.txt
grpcio
grpcio-tools
grpc-stubs
types-protobuf
protobuf
click
